% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_biome.R
\name{load_biome}
\alias{load_biome}
\title{Load a curated biome phyloseq object}
\usage{
load_biome(biome_id)
}
\arguments{
\item{biome_id}{A biome ID from \code{list_biomes()}.}
}
\value{
A phyloseq object.
}
\description{
Load a curated biome phyloseq object
}
