% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{package}
\name{metaprodb-package}
\alias{metaprodb}
\alias{metaprodb-package}
\title{metaprodb: Framework for Biome-Informed Metaproteomic Database Construction}
\description{
Tools for selecting taxa from curated biome objects or user-provided inputs, generating manifests, and orchestrating construction of biome-informed metaproteomic protein databases.
}
\author{
\strong{Maintainer}: Muzaffer Arikan \email{muzafferarikan@istanbul.edu.tr}

}
\keyword{internal}
