% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genus_index.R
\name{read_genus_index}
\alias{read_genus_index}
\title{Read genus genome resource index}
\usage{
read_genus_index(path)
}
\arguments{
\item{path}{Path to genus resource index TSV file.}
}
\value{
A tibble describing available genus genome resources.
}
\description{
Read genus genome resource index
}
