% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_all_biomes.R
\name{summarize_all_biomes}
\alias{summarize_all_biomes}
\title{Summarize all curated biomes}
\usage{
summarize_all_biomes()
}
\value{
A tibble with one row per biome.
}
\description{
Runs \code{summarize_biome()} on all bundled biomes.
}
