% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_build_plan.R
\name{summarize_build_plan}
\alias{summarize_build_plan}
\title{Summarize a database build plan}
\usage{
summarize_build_plan(build_plan_tbl)
}
\arguments{
\item{build_plan_tbl}{Output of \code{plan_database_build()}.}
}
\value{
A one-row tibble summarizing build readiness.
}
\description{
Summarize a database build plan
}
