% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_build_plan_tsv.R
\name{write_build_plan_tsv}
\alias{write_build_plan_tsv}
\title{Write a database build plan to TSV}
\usage{
write_build_plan_tsv(build_plan_tbl, path)
}
\arguments{
\item{build_plan_tbl}{Output of \code{plan_database_build()}.}

\item{path}{Output TSV path.}
}
\value{
Invisibly returns \code{path}.
}
\description{
Write a database build plan to TSV
}
