% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_snakemake_config.R
\name{write_snakemake_config}
\alias{write_snakemake_config}
\title{Write a Snakemake config file for database building}
\usage{
write_snakemake_config(
  build_plan_path,
  final_database,
  resources_dir,
  path,
  host_proteome = NULL
)
}
\arguments{
\item{build_plan_path}{Path to build plan TSV.}

\item{final_database}{Path to final output FASTA.}

\item{resources_dir}{Directory containing genome zip resources.}

\item{path}{Output YAML path.}

\item{host_proteome}{Optional path to a host protein FASTA file.}
}
\value{
Invisibly returns \code{path}.
}
\description{
Write a Snakemake config file for database building
}
